requirements.txt
Django~=4.0.1
djangorestframework~=3.13.1
python-dateutil~=2.8.2